feat: 리뷰 목록 페이지에  groupAccessCode 전역 상태 적용
